Absorb perfgate-error into perfgate-types
