+++ crates/perfgate-error/Cargo.toml
@@ -7,19 +7,14 @@ license.workspace = true
 repository.workspace = true
 homepage.workspace = true
 authors.workspace = true
-description = "Unified error types for perfgate"
+description = "Compatibility wrapper for perfgate error contract types"
 readme = "README.md"
 documentation = "https://docs.rs/perfgate-error"
 keywords = ["error", "error-handling", "performance", "benchmarking"]
 categories = ["development-tools"]
 
 [dependencies]
-serde_json.workspace = true
-thiserror.workspace = true
-toml.workspace = true
-
-[dev-dependencies]
-proptest = "1.7"
+perfgate-types.workspace = true
 
 [lints]
 workspace = true
