+++ crates/perfgate-types/Cargo.toml
@@ -22,13 +22,12 @@ serde.workspace = true
 serde_json.workspace = true
 schemars.workspace = true
 thiserror.workspace = true
-perfgate-error.workspace = true
+toml.workspace = true
 arbitrary = { version = "1", features = ["derive"], optional = true }
 
 [dev-dependencies]
 insta.workspace = true
 tempfile.workspace = true
-toml.workspace = true
 proptest = "1"
 proptest-derive = "0.4"
