+++ docs/WORKSPACE.md
@@ -6,7 +6,7 @@ This file is automatically generated by `cargo run -p xtask -- docs-sync`.
 
 | Crate | Description | Kill Rate Target |
 |-------|-------------|------------------|
-| `perfgate-error` | Unified error types for error propagation | 100% |
+| `perfgate-error` | Compatibility wrapper for perfgate_types::error | 100% |
 | `perfgate-sha256` | Minimal SHA-256 implementation (no_std compatible) | 100% |
 | `perfgate-host-detect` | Host mismatch detection for CI noise reduction | 100% |
 | `perfgate-budget` | Budget evaluation logic for performance thresholds | 100% |
@@ -34,7 +34,7 @@ This file is automatically generated by `cargo run -p xtask -- docs-sync`.
 
 ```mermaid
 graph TD
-  error[perfgate-error] --> types[perfgate-types]
+  error[perfgate-error compatibility wrapper] --> types[perfgate-types]
   sha[perfgate-sha256] --> types
   domain --> stats[perfgate-domain::stats]
   types --> val[perfgate-types::validation]
